Clean-up a bit
