+++ netlify.toml
@@ -5,15 +5,15 @@
 
 [[redirects]]
   from = "/catalog/Authors"
-  to = "/.netlify/functions/server3/Authors"
+  to = "/.netlify/functions/catalog/Authors"
   status = 200
 
 [[redirects]]
   from = "/catalog/Books"
-  to = "/.netlify/functions/server3/Books"
+  to = "/.netlify/functions/catalog/Books"
   status = 200
 
 [[redirects]]
   from = "/catalog/$metadata"
-  to = "/.netlify/functions/server3/$metadata"
+  to = "/.netlify/functions/catalog/$metadata"
   status = 200
